requirements.txt
requests
yt-dlp
instaloader
ffmpeg-python
python-telegram-bot
python-dotenv
